First test with back to back ALU instructions
This commit implements a very basic testbench that which reads from an
instruction file and supplies the instructions upon request to the IF
unit. The IF Unit is primarily the same as in RI5CY but does not
implement control flow changes of any kind at the moment. It is
furthermore more decoupled from the rest of the pipeline as it is done
in all major processors:

The processor requests instruction from the IF stage and acknowledges
them. At the moment the processor only supports 64bit register register
instructions, NOP and load immediate (ADDI). Results are written back by
the simple commit unit.